Move the C API to a new crate
The crate stracciatella becomes a static rust lib (rlib).
The new crate stracciatella_c_api is a static C lib.

The test in module c::librarydb was removed.
It will probably be recreated later as C++ tests.
