+++ rust/CMakeLists.txt
@@ -21,9 +21,9 @@ if (MINGW)
     set(LIBSTRACCIATELLA_RUSTFLAGS "${LIBSTRACCIATELLA_RUSTFLAGS} -C linker=${CMAKE_C_COMPILER}")
 endif()
 
-set(LIBSTRACCIATELLA_NAME "${CMAKE_STATIC_LIBRARY_PREFIX}stracciatella${CMAKE_STATIC_LIBRARY_SUFFIX}")
+set(LIBSTRACCIATELLA_NAME "${CMAKE_STATIC_LIBRARY_PREFIX}stracciatella_c_api${CMAKE_STATIC_LIBRARY_SUFFIX}")
 if (MINGW)
-    set(LIBSTRACCIATELLA_NAME "stracciatella.lib")
+    set(LIBSTRACCIATELLA_NAME "stracciatella_c_api.lib")
 endif()
 if(LIBSTRACCIATELLA_TARGET)
     set(LIBSTRACCIATELLA_BUILD_SWITCHES ${LIBSTRACCIATELLA_BUILD_SWITCHES} "--target=${LIBSTRACCIATELLA_TARGET}")
