+++ rust/stracciatella/src/librarydb.rs
@@ -20,9 +20,9 @@
 //!
 //! # FFI
 //!
-//! [`stracciatella::c::librarydb`] contains a C interface for this module.
+//! [`stracciatella_c_api::c::librarydb`] contains a C interface for this module.
 //!
-//! [`stracciatella::c::librarydb`]: ../c/librarydb/index.html
+//! [`stracciatella_c_api::c::librarydb`]: ../../stracciatella_c_api/c/librarydb/index.html
 
 use std::cmp;
 use std::collections::HashSet;
