+++ rust/stracciatella_c_api/src/lib.rs
@@ -1,0 +1,5 @@
+//! This crate contains the C API of the [stracciatella] crate.
+//!
+//! [stracciatella]: ../stracciatella/index.html
+
+pub mod c;
